Moved vmec_from_surf() function to helper file
